code improvement by eclipse suggestions
git-svn-id: file:///home/SUBVERSION/jas/trunk@573 eafcb37d-f50b-0410-b8f7-d4cb9effa109

+++ poly/ExpVector.java
@@ -96,7 +96,7 @@ public class ExpVector implements Cloneable, Serializable {
             int length = exps.size();
             val = new long[ length ];
             for ( int j = 0; j < length; j++ ) {
-                val[j] = ((Long)exps.elementAt(j)).longValue();
+                val[j] = exps.elementAt(j).longValue();
             }
         } else {
         // not implemented
